--- src/main/java/com/pinHouse/server/core/config/SwaggerConfig.java
@@ -1,36 +1,0 @@
-package com.pinHouse.server.core.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        String jwt = "JWT";
-        SecurityRequirement securityRequirement = new SecurityRequirement().addList(jwt);
-        Components components = new Components().addSecuritySchemes(jwt, new SecurityScheme()
-                .name(jwt)
-                .type(SecurityScheme.Type.HTTP)
-                .scheme("Bearer")
-                .bearerFormat("JWT")
-        );
-        return new OpenAPI()
-                .components(components)
-                .info(apiInfo())
-                .addSecurityItem(securityRequirement);
-    }
-
-    private Info apiInfo() {
-        return new Info()
-                .version("1.0")
-                .title("집터 API")
-                .description("집터 개발 서버의 API 입니다");
-    }
-}
